requirements.txt
streamlit==1.23.1
PyMuPDF==1.22.5
sentence-transformers==2.2.2
faiss-cpu==1.12.0
transformers==4.32.0
torch==2.8.0
python-docx
python-pptx
exceptions